Allow selecting instances directly from the command line
This commit adds a new --match (or -m) command line option that allows
passing a fuzzy string to match against all the columns of an instance
in the instance table. The table will be filtered to show only the rows
of the matching instances. If only a single instance matches, the prompt
is skipped and that instance is selected automatically.